pyproject.toml
[tool.poetry]
name = "swe-hw1-backend"
version = "0.1.0"
description = ""
authors = ["mythicalbadger <jussy12.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
black = "^23.12.1"
flake8 = "^7.0.0"
sqlmodel = "^0.0.14"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.1"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
